requirements.txt
Django==2.0.3
#pandas==1.1.4
#numpy==1.19.4
Pillow==5.0.0
django-cleanup==5.1.0
